Backend/pyproject.toml
[tool.poetry]
name = "ietp-back"
version = "0.1.0"
description = ""
authors = ["minda"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
django-tailwind = {extras = ["reload"], version = "^3.8.0"}
django-filter = "^24.3"
heroicons = {extras = ["django"], version = "^2.9.0"}
faker = "^33.1.0"
django-htmx = "^1.21.0"
django-extensions = "^3.2.3"
graphviz = "^0.20.3"
pygraphviz = "^1.14"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
